Harvest incrementally with recurring cron jobs
Record the ingestion date of the last successfully gathered
product and use it as the start date of the next job. Limit
the number of datasets harvested in each job. We receive
feedback sooner if there are errors and there is no need to save
failed queries since new jobs can pick up directly from where
the last job stopped. It is now possible to harvest all datasets
starting from the earliest without creating a huge job or manually
updating the start and end dates before each job runs.