Resolve [#31]
This commit removes all the unused/never executed/unnecessary code
related to the issue.

There's still no way to use the harvester to update (or delete)
a previously harvested dataset, but that's something we'll need
to address in a different issue.

Currently attempts to reharvest an existing dataset fail with a
validation error when package_create is called (because the name
is already in use).