Streamline check and create workflow
Rather than a separate method for each type of harvester we now
have one method that for all of them with a few helper methods
to create the appropriate metadata based on the harvester type.

The point isn't to be clever or elegant--the point is to reduce
duplicated code as much as possible. And not only for clarity and
understandability, but in order to reduce errors. There were
errors in previous versions that would have been easier to catch
if there was only one place to look.