Add functional test for CMEMS harvester
The test runs the harvester from start to finish and verifies that
it works as expected (except for the FTP request, which we have to
mock).

We should parmeterize the test for all four harvester types,
probably using the parameterized library, but I'm running into
errors trying to get it set up. (Testing each harvester type
individually works, but trying to test them all using the
parameterized decorator does not.)