Remove duplicated try/except for CMEMS requests
The removed code was intended to catch exceptions while making
requests to the FTP server, but 1) the _crawl_urls_ftp() method
already handles those exceptions, 2) the exceptions that would
be caught would be reported inaccurately since the value of url
would always be 'dummy'. If we do need to catch errors here,
we should think about how to handle them--we might not want to
abort the process (return None) in every case.