Make source_config attribute in import stage too
When the test runner runs a harvester, it runs through all the
stages at once and only instantiates one harvest object. That
means the source_config attribute created in the gather_stage still
exists during the import stage. When the harvester runs "normally",
the stages run separately and the harvest object is initiated anew
each time, so we need to create the source_config attribute in each
stage that requires it.